+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2019.
 * Add support for downloading user data.
 * Addons are now automatically triggered upon installation.
 * Improved instructions for resolving merge conflicts.
+* Cleanup addon is now compatible with app store metadata translations.
 
 weblate 3.5.1
 -------------